Adding (optional) file name as a prefix to the print statement.
This commit is also a somewhat big refactoring of the code. Now we only have
"multi-var-formatting-fn", which makes it easier to finely control the finer
points of the output string.

We also include the prepare-output code in the generated program, so we can have
access to its wrapping context (i.e., modifiers).